implement delete of players

+++ Backend/DataLayer/Sqlite3/SellPlayer.php
@@ -9,6 +9,14 @@ class SellPlayer extends Player implements WriteablePlayer
         if (!$this->id) {
             throw new \Exception('Internal error: player is uninitialized, cannot delete it');
         }
+        if (!$this->exists()) {
+            return;
+        }
+        $data = $this->db->query("SELECT * FROM player WHERE id='" . $this->db->escapeString($this->id) . "'");
+        $row = $data->fetchArray(SQLITE3_ASSOC);
+        if ($this->code !== $row['createstamp']) {
+            throw new \Exception("Error: code does not match");
+        }
         $this->db->exec('BEGIN');
         try {
             $this->db->exec("DELETE FROM player WHERE id='" . $this->db->escapeString($this->id) . "'");
